% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{dim.Tensor}
\alias{dim.Tensor}
\alias{dim-Tensor}
\title{Tensor dim method}
\usage{
\method{dim}{Tensor}(x)
}
\arguments{
\item{x}{A Tensor object}
}
\value{
A length 2 vector of the number of rows and columns respectively.
}
\description{
Retrieve dimension of object
}
\author{
Charles Determan Jr.
}
