% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializers.R
\name{get_initializer}
\alias{get_initializer}
\title{Retrive Initializer Function}
\usage{
get_initializer(initializer)
}
\arguments{
\item{initializer}{Character scalar defining initializer Tensor to return}
}
\description{
Returns an initializer function to
generate a Tensor object.
}
