% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializers.R
\docType{data}
\name{Initializer}
\alias{Initializer}
\alias{Zeros}
\alias{Ones}
\alias{Constant}
\alias{RandomNormal}
\title{Initializer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Initializer

Zeros

Ones

Constant

RandomNormal
}
\value{
Object of class relevant to defined backend
}
\description{
Initializer for a tensor
}
\section{Usage}{

\preformatted{
z = Zeros$new(shape = c(4,4))
o = Ones$new(shape = c(4,4))
c = Constant$new(shape = c(4,4), constant = 2)
r = RandomNormal$new(shape = c(4,4))
}
}

\section{Public Methods}{

\describe{
  \item{\code{compute()}}{Evaluate Initializer}
}
}

\author{
Charles Determan Jr.
}
\keyword{datasets}
