% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{matmult}
\alias{matmult}
\title{Tensor Matrix Multiplication}
\usage{
matmult(x, y)
}
\arguments{
\item{x}{base R or Tensor object}

\item{y}{base R or Tensor object}
}
\value{
The \code{x} Tensor with the updated internal graph (invisible)
}
\description{
A higher level interface that may
be more comfortable that using the internal \code{.dot}
method for dot products.
}
