% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{mean.Tensor}
\alias{mean.Tensor}
\title{#' @export
mean.Tensor <- function(x, trim = 0, na.rm = FALSE){
  x$mean(trim, na.rm)
  return(invisible(x))
}}
\usage{
\method{mean}{Tensor}(x, trim = 0, na.rm = FALSE, ...)
}
\arguments{
\item{x}{A Tensor object}

\item{trim}{the fraction (0 to 0.5) of observations to be trimmed from
each end of x before the mean is computed. Values of trim outside that
range are taken as the nearest endpoint.}

\item{na.rm}{a logical value indicating whether NA values should be
stripped before the computation proceeds.}

\item{...}{further arguments to or from other methods}
}
\description{
#' @export
mean.Tensor <- function(x, trim = 0, na.rm = FALSE){
  x$mean(trim, na.rm)
  return(invisible(x))
}
}
