% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{nrow}
\alias{nrow}
\alias{ncol}
\title{The Number of Rows/Columns of a Tensor matrix}
\usage{
nrow(x)

ncol(x)
}
\arguments{
\item{x}{A Tensor object}
}
\value{
An integer of length 1
}
\description{
\code{nrow} and \code{ncol} return the number of rows or columns
present in \code{x} respectively.
}
\author{
Charles Determan Jr.
}
