% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{setBackend}
\alias{setBackend}
\title{Set Backend Type}
\usage{
setBackend(backend)
}
\arguments{
\item{backend}{A character scalar defining the backend (e.g.)}
}
\description{
Select type of tensor (i.e. matrix, vector, etc.) to
have lazytensor to use internally
}
