% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tensor.R
\docType{data}
\name{Tensor}
\alias{Tensor}
\alias{Placeholder}
\alias{variable}
\alias{constant}
\title{Tensor}
\format{\code{\link{R6Class}} object}
\usage{
Tensor

Placeholder

variable

constant
}
\value{
Object of \code{\link{R6Class}} with methods for symbolic operations
}
\description{
Tensor class providing methods for chaining operations to be
evaluated lazily.
}
\section{Fields}{

\describe{
\item{\code{tensor}}{Stores the tensor object}

\item{\code{name}}{The name of the Tensor object}

\item{\code{graph}}{The graph containing the registered Nodes and
Operations to be evaluated}
}}

\section{Usage}{

\preformatted{
rmat = matrix(rnorm(16), 4)
a = Tensor$new(rmat)
init = Tensor$new("RandomNormal", shape = c(4,4))
v = variable$new(rmat)
c = constant$new(rmat)
}
}

\section{Public Methods}{

\describe{
  \item{\code{.dot(name = NA)}}{Dot product of two Tensors}
  \item{\code{.add(name = NA)}}{Addition of two Tensors}
  \item{\code{.sub(name = NA)}}{Subtraction of two Tensors}
  \item{\code{.mult(name = NA)}}{Elementwise multiplication of two Tensors}
  \item{\code{.div(name = NA)}}{Elementwise division of two Tensors}
  \item{\code{.eq(name = NA)}}{Elementwise equality of Tensors}
  \item{\code{.neq(name = NA)}}{Elementwise inequality of Tensors}
  \item{\code{.gte(name = NA)}}{Elementwise >= of Tensors}
  \item{\code{.gt(name = NA)}}{Elementwise > of Tensors}
  \item{\code{.lte(name = NA)}}{Elementwise <= of Tensors}
  \item{\code{.lt(name = NA)}}{Elementwise < of Tensors}
  \item{\code{pow(val,name = NA)}}{Power (\code{^}) of Tensor elements}
  \item{\code{log(base = exp(1))(name = NA)}}{Natural log of Tensor elements}
  \item{\code{log10(name = NA)}}{Log base 10 of Tensor elements}
  \item{\code{log1p(name = NA)}}{\code{log(1+x)} of Tensor elements}
  \item{\code{log2(name = NA)}}{Log base 2 of Tensor elements}
  \item{\code{exp(name = NA)}}{Exponential of Tensor elements}
  \item{\code{expm1(name = NA)}}{\code{exp(x) - 1} of Tensor elements}
  \item{\code{sin(name = NA)}}{sin of Tensor elements}
  \item{\code{asin(name = NA)}}{asin of Tensor elements}
  \item{\code{sinh(name = NA)}}{sinh of Tensor elements}
  \item{\code{cos(name = NA)}}{cos of Tensor elements}
  \item{\code{acos(name = NA)}}{acos of Tensor elements}
  \item{\code{cosh(name = NA)}}{cosh of Tensor elements}
  \item{\code{tan(name = NA)}}{tan of Tensor elements}
  \item{\code{atan(name = NA)}}{atan of Tensor elements}
  \item{\code{tanh(name = NA)}}{tanh of Tensor elements}
  \item{\code{max(name = NA)}}{max of Tensor elements}
  \item{\code{min(name = NA)}}{min of Tensor elements}
  \item{\code{mean(name = NA)}}{mean of Tensor elements}
  \item{\code{abs(name = NA)}}{abs of Tensor elements}
  \item{\code{round(digits, name = NA)}}{round of Tensor elements}
  \item{\code{floor(name = NA)}}{floor of Tensor elements}
  \item{\code{ceiling(name = NA)}}{ceiling of Tensor elements}
  \item{\code{sqrt(name = NA)}}{sqrt of Tensor elements}
  \item{\code{sign(name = NA)}}{sign of Tensor elements}
  \item{\code{sum(name = NA)}}{sum of Tensor elements}
  \item{\code{cumsum(name = NA)}}{cumulative sum of Tensor elements}
  \item{\code{prod(name = NA)}}{product of Tensor elements}
  \item{\code{cumprod(name = NA)}}{cumulative product of Tensor elements}
  \item{\code{compute(feed_list)}}{Compute/Evaluate the Tensor graph}
  \item{\code{chain(f)(name = NA)}}{Process a function to be evaluated within the Tensor graph}
  \item{\code{drop(idx = NA, name = NA)}}{Drop a previously added operation from the Tensor graph}
  \item{\code{has_history}}{returns boolean indicating of additional Tensor involved in graph}
}
}

\section{Details}{

\code{Tensor} is the base class of this package (based upon R6).  The
\code{Placeholder}, \code{variable}, and \code{constant} are all child
class of the \code{Tensor} class.

There are only minor differences with the \code{variable} and \code{constant}
child classes.  \code{variable} instances have their \code{shape} fixed.  The
internal data values may be changed but the shape of the Tensor is not allowed.
As can be assumed, the \code{constant} child class does not allow any change to
the underlying Tensor or the Tensor shape.

The \code{Placeholder} class is unique in that it contains NO underlying data.
The class simply contains the shape of the intended Tensor.  All chained methods
are still applicable.  The user is recommended to provide a \code{name} to the
initialized \code{Placeholder} as when the final \code{compute} method is called,
the operations will look for a passed named list element in the
\code{compute{feed_list = list()}} call.  This allows the user to provide alternate
datasets to a previously prototyped process.  This mirrors the functionality of
libraries such as Tensorflow.
}

\author{
Charles Determan Jr.
}
\keyword{datasets}
