% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{mean.Tensor}
\alias{mean.Tensor}
\title{#' @export
mean.Tensor <- function(x, trim = 0, na.rm = FALSE){
  x$mean(trim, na.rm)
  return(invisible(x))
}}
\usage{
\method{mean}{Tensor}(x, ...)
}
\description{
#' @export
mean.Tensor <- function(x, trim = 0, na.rm = FALSE){
  x$mean(trim, na.rm)
  return(invisible(x))
}
}
